# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Wine App API
  version: 0.1.0
  description: Current API endpoints for health, authentication, wines, reviews, stats, and wine photos. Web clients continue to use the Symfony session cookie; mobile clients should prefer bearer JWT auth.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://dev-api.example.com
    description: Shared development environment placeholder
  - url: https://staging-api.example.com
    description: Staging environment placeholder
  - url: https://api.example.com
    description: Production environment placeholder
paths:
...
  /api/wines/{wineId}/reviews:
    post:
      summary: Create review
      operationId: postWineReviewsCreate
      tags: [Review]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: wineId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReviewResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Review already exists for this user and wine
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/reviews:
    get:
      summary: List reviews (paginated)
      operationId: getReviewsList
      tags: [Review]
      parameters:
        - in: query
          name: page
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: user_id
          description: Filter by user id. Use `me` to restrict to the authenticated user when available.
          schema:
            oneOf:
              - type: integer
                format: int64
              - type: string
                enum: [me]
        - in: query
          name: sort_by
          description: Sort field
          schema:
            type: string
            enum: [score, name, do]
            default: score
        - in: query
          name: sort_dir
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated review list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '400':
          description: Invalid query params
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/wines/{wineId}/reviews/{id}:
    get:
      summary: Get review by id
      operationId: getWineReviewById
      tags: [Review]
      parameters:
        - in: path
          name: wineId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Review found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update review by id
      operationId: putWineReviewById
      tags: [Review]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: wineId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '204':
          description: Review updated
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (review owned by another user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete review by id
      operationId: deleteWineReviewById
      tags: [Review]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: wineId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Review deleted
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (review owned by another user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: PHPSESSID
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Invalid credentials.
    ReviewBulletValue:
      type: string
      enum: [fruity, floral, mineral, oak_forward, powerful]
    Review:
      type: object
      required: [id, user_id, wine_id, aroma, appearance, palate_entry, body, persistence, bullets]
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        wine_id:
          type: integer
          format: int64
        score:
          type: integer
          nullable: true
          minimum: 0
          maximum: 100
        aroma:
          type: integer
          minimum: 0
          maximum: 10
        appearance:
          type: integer
          minimum: 0
          maximum: 10
        palate_entry:
          type: integer
          minimum: 0
          maximum: 10
        body:
          type: integer
          minimum: 0
          maximum: 10
        persistence:
          type: integer
          minimum: 0
          maximum: 10
        bullets:
          type: array
          items:
            $ref: '#/components/schemas/ReviewBulletValue'
        created_at:
          type: string
          format: date-time
          nullable: true
    CreateReviewRequest:
      type: object
      required: [aroma, appearance, palate_entry, body, persistence]
      properties:
        created_at:
          type: string
          nullable: true
          description: Review date. Accepts `YYYY-MM-DD` or ISO-8601 date-time.
          example: '2025-12-24'
        score:
          type: integer
          nullable: true
          minimum: 0
          maximum: 100
        aroma:
          type: integer
          minimum: 0
          maximum: 10
        appearance:
          type: integer
          minimum: 0
          maximum: 10
        palate_entry:
          type: integer
          minimum: 0
          maximum: 10
        body:
          type: integer
          minimum: 0
          maximum: 10
        persistence:
          type: integer
          minimum: 0
          maximum: 10
        bullets:
          type: array
          items:
            $ref: '#/components/schemas/ReviewBulletValue'
    UpdateReviewRequest:
      type: object
      required: [aroma, appearance, palate_entry, body, persistence]
      properties:
        created_at:
          type: string
          nullable: true
          description: Review date. Accepts `YYYY-MM-DD` or ISO-8601 date-time.
          example: '2025-12-24'
        score:
          type: integer
          nullable: true
          minimum: 0
          maximum: 100
        aroma:
          type: integer
          minimum: 0
          maximum: 10
        appearance:
          type: integer
          minimum: 0
          maximum: 10
        palate_entry:
          type: integer
          minimum: 0
          maximum: 10
        body:
          type: integer
          minimum: 0
          maximum: 10
        persistence:
          type: integer
          minimum: 0
          maximum: 10
        bullets:
          type: array
          items:
            $ref: '#/components/schemas/ReviewBulletValue'
    CreateReviewResponse:
      type: object
      required: [review]
      properties:
        review:
          type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
    ReviewResponse:
      type: object
      required: [review]
      properties:
        review:
          $ref: '#/components/schemas/Review'
    ReviewListUser:
      type: object
      required: [id, name, lastname]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        lastname:
          type: string
    ReviewListDo:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    ReviewListWine:
      type: object
      required: [id, name, do]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        do:
          allOf:
            - $ref: '#/components/schemas/ReviewListDo'
          nullable: true
    ReviewListItem:
      type: object
      required: [id, user, wine, score, aroma, appearance, palate_entry, body, persistence, bullets, created_at]
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/ReviewListUser'
        wine:
          $ref: '#/components/schemas/ReviewListWine'
        score:
          type: integer
          nullable: true
          minimum: 0
          maximum: 100
        aroma:
          type: integer
          minimum: 0
          maximum: 10
        appearance:
          type: integer
          minimum: 0
          maximum: 10
        palate_entry:
          type: integer
          minimum: 0
          maximum: 10
        body:
          type: integer
          minimum: 0
          maximum: 10
        persistence:
          type: integer
          minimum: 0
          maximum: 10
        bullets:
          type: array
          items:
            $ref: '#/components/schemas/ReviewBulletValue'
        created_at:
          type: string
          format: date-time
    ReviewListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewListItem'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      required: [page, limit, total_items, total_pages, has_next, has_prev]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
...